% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{addStudyTimes.lst}
\alias{addStudyTimes.lst}
\title{Function to add study week and day columns (integers) to a list of data.tables}
\usage{
addStudyTimes.lst(
  lst,
  dt.dates,
  col.time = "StartTime",
  col.datesId = "record_id",
  col.datesStudyStartDate = "study_start_date",
  col.datesStudyWeek = "week",
  col.datesStudyWeekDate = "start_date",
  col.study.day = "study.day",
  col.study.week = "study.week",
  num.week.min = -1000,
  num.week.max = 1000
)
}
\arguments{
\item{dt.dates}{data.table with study date ranges including study start date and weeks of interest}

\item{col.time}{column with time in dt (POSIX format expected)}

\item{col.datesId}{column with subject ids in dt.dates}

\item{col.datesStudyStartDate}{column with study start date in dt.dates}

\item{col.datesStudyWeek}{column with week numbers in dt.dates}

\item{col.datesStudyWeekDate}{column with week dates in dt.dates}

\item{col.study.day}{column to add to dt with study days}

\item{col.study.week}{column to add to dt with study weeks}

\item{num.week.min}{minimum week value to use (set very low by default to include all data)}

\item{dt}{data.table}

\item{week.max}{maximum week value to use (set very high by default to include all data)}
}
\value{
list of data.tables with dt + col.study.day and col.study.week
}
\description{
Function to add study week and day columns (integers) to a list of data.tables
}
