% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{aggregateList.dt}
\alias{aggregateList.dt}
\title{Aggregate List}
\usage{
aggregateList.dt(lst, col.id, col.measure, vec.range, fn.lstAggregate,
  str.variable, str.value, fn.rangeAggregate = median)
}
\arguments{
\item{lst}{list with data}

\item{col.id}{id in list (ex. study day)}

\item{col.measure}{column to measure in list}

\item{vec.range}{vector with range of id values over which to aggregate}

\item{fn.lstAggregate}{function to aggregate list data}

\item{str.variable}{name for id column in new table after aggregation over all col.id in vec.idRange}

\item{str.value}{name for value column in new table}

\item{fn.rangeAggregate}{function to aggregate data in new table (ex. median of days in the specified date range). Default: median}
}
\value{
data.table with aggregated data
}
\description{
Helper function to aggregate list data into columns that can later be merged into a monolithic table
}
