% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_TApos.R
\name{calc_TApos}
\alias{calc_TApos}
\title{Map TA sites onto corresponding genes}
\usage{
calc_TApos(tally, genelist)
}
\arguments{
\item{tally}{raw tally.txt file that imported from \code{import_raw_tally}
function.}

\item{internal}{data in the package, strain-specific, containing gene name,
TA site positions of each gene.}
}
\value{
numeric vector of positions.
}
\description{
This function is part of the \code{tallyprepfun} function, which pre-process
raw tally file for analysis. In this step, the function uses raw tally .txt
file (imported from function \code{import_raw_tally})
}
