% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcparafun.R
\name{calcparafun}
\alias{calcparafun}
\title{Calculate the parameters!  (I don't yet know what they are, but I suspect
they are the mu and sigma in the text of the paper.}
\usage{
calcparafun(strain, usable_tally_list, save_location, rep_time)
}
\arguments{
\item{strain}{Bacterial strain to examine.}

\item{usable_tally_list}{Presumably the parsed tally.txt}

\item{save_location}{Where to save the result of this function.}

\item{rep_time}{How many times to repeat the parameter collection, I presume.}
}
\value{
List containing the strain/experiment specific parameters.
}
\description{
Calculate the parameters!  (I don't yet know what they are, but I suspect
they are the mu and sigma in the text of the paper.
}
