% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callessfun.R
\name{callessfun}
\alias{callessfun}
\title{Perform the actual calling of essentiality.}
\usage{
callessfun(file_location, usable_tally_list, parameter_list)
}
\arguments{
\item{file_location}{What to parse for the calls.}

\item{usable_tally_list}{Subset of the tally.txt to call.}

\item{parameter_list}{The parameters from calcparafun().}
}
\value{
Essentiality!
}
\description{
Perform the actual calling of essentiality.
}
