% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecog_functions.r
\name{correctEcog.num}
\alias{correctEcog.num}
\title{Correct ECOG (numeric)}
\usage{
correctEcog.num(num.ecog, num.angle, num.cutoffLow = 20,
  num.cutoffModerate = 30, num.cutoffHigh = 50)
}
\arguments{
\item{num.ecog}{ECOG score}

\item{num.angle}{angle from linear model of HR vs. steps}

\item{num.cutoffLow}{lowest angle to use for ECOG correction. Default: 20}

\item{num.cutoffModerate}{lowest angle corresponding to intense activity. Default: 30}

\item{num.cutoffHigh}{highest angle to use for ECOG correction. Default: 50}
}
\value{
numeric corrected ECOG score
}
\description{
Function to correct ECOG for HR response using an angle derived from linear modeling of HR vs steps
}
\details{
Default ECOG HR correction:
  angle > 50: 0
  angle 30-50: -2
  angle 20-30: -1
  angle < 20: 0
}
