% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecog_functions.r
\name{correctEcog.pct}
\alias{correctEcog.pct}
\title{Correct ECOG (percentage)}
\usage{
correctEcog.pct(num.ecog, pct.veryActive, pct.fairlyActive,
  num.cutoffVery = 0.05, num.cutoffFairly = 0.1)
}
\arguments{
\item{num.ecog}{ECOG score}

\item{pct.veryActive}{percentage of 'very active' time}

\item{num.cutoffVery}{lowest percentage to use 'very active' ECOG correction. Default: 0.05}

\item{num.cutoffFairly}{lowest percentage to use 'fairly active' ECOG correction. Default: 0.10}

\item{num.pctFairlyActive}{percentage of 'fairly active' time}
}
\value{
numeric corrected ECOG score
}
\description{
Function to correct ECOG for percentage of time spent being 'fairly' or 'very' active (fitbit variables).
}
\details{
Default ECOG correction:
  fairly + very active < %10: 0
  very active >= 5%: -2
  fairly active >= 10%: -1
}
