% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{createLabels.dt}
\alias{createLabels.dt}
\title{Create labels}
\usage{
createLabels.dt(dir.files, str.filePattern, str.idPattern = "^([^_]+)_.*")
}
\arguments{
\item{dir.files}{directory containing the files}

\item{str.filePattern}{file name pattern}

\item{str.idPattern}{regular expression to identify sample id in filename. Default: '^(\link{^_}+)_.*'}
}
\value{
data.table with dir, file, and label (id)
}
\description{
Function to create a sample key / label table
}
