% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denote_coreTA.R
\name{denote_coreTA}
\alias{denote_coreTA}
\title{A quick and dirty search for core TAs.}
\usage{
denote_coreTA(tally, bp = 50)
}
\arguments{
\item{tally}{Parsed tally data.}

\item{bp}{Minimum number of bps!}
}
\value{
a new tally.
}
\description{
A quick and dirty search for core TAs.
}
