% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{filterDates.dt}
\alias{filterDates.dt}
\title{Filter dates}
\usage{
filterDates.dt(
  dt,
  date.start,
  days.up = 7,
  days.down = 7,
  col.date = "date",
  col.winDate = "day.window",
  str.format = "\%m-\%d-\%Y"
)
}
\arguments{
\item{dt}{data.table with a date column}

\item{date.start}{date to filter around}

\item{days.up}{days prior to date.start to include.}

\item{days.down}{days after to date.start to include.}

\item{col.date}{name of the date column in dt.}

\item{col.winDate}{name of the column to create with the relative number of days}

\item{str.format}{date string format to use for any non-date formatted columns.}
}
\value{
data.table subset with the dates of interest and an added column of days relative to date.start
}
\description{
Function to filter a data table around a specific date
}
