% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_homo.R
\name{find_homo}
\alias{find_homo}
\title{Find TAs which are homologous, to what I am not certain.}
\usage{
find_homo(tally, homofile)
}
\arguments{
\item{tally}{Parsed tally.txt}

\item{homofile}{File containing homology information.}
}
\description{
Find TAs which are homologous, to what I am not certain.
}
