% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\docType{package}
\name{fitness}
\alias{fitness}
\title{Working with Fitness Landscapes}
\description{
A helpful library for useRs studying fitness landscapes. `fitness`
implements popular models for fitness landscapes (e.g. NK model,
rough Mt. Fuji) and associated functionality (e.g. optima identification).
}
