% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessRun_function.R
\name{fitnessRun}
\alias{fitnessRun}
\title{Main function for the FiTnEss library.}
\usage{
fitnessRun(strain, file_location, save_location = "fitness_output.xlsx",
  repeat_time = 3)
}
\arguments{
\item{strain}{Name of the bacterial strain being analyzed.}

\item{file_location}{Filename(s) of the numbers of reads by position.}

\item{save_location}{Output .xlsx filename.}

\item{repeat_time}{Number of times to repeat the parameter calculations.}
}
\value{
An excel file of the results.
}
\description{
This function calls the various workers inside FiTnEss.
}
