% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{generate_nk}
\alias{generate_nk}
\title{Generate Fitness Landscape using NK Model}
\usage{
generate_nk(n, k)
}
\arguments{
\item{n}{corresponds to the N parameter in the NK model}

\item{k}{corresponds to the K parameter in the NK model}
}
\value{
data frame containing landscape generated via NK model
}
\description{
A fitness landscape generated by the NK model.
The landscape is returned as a two-column data frame,
where the first column contains the coordinates of
the spot (class `character`) and the second column contains
the fitness value (class `numeric`). For more information
on NK models, see Kauffman & Weinberger (1989)
<doi:10.1016/s0022-5193(89)80019>.
}
