% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{generate_rmf}
\alias{generate_rmf}
\title{Generate Fitness Landscape using RMF model}
\usage{
generate_rmf(max.seq, c, eta = function() rnorm(1))
}
\arguments{
\item{max.seq}{genotype for the peak of the RMF model}

\item{c}{corresponds to the `c` parameter in the RMJ model}

\item{eta}{function that returns outputs of a random variable;
default value is `function() rnorm(1)`.}
}
\value{
data frame containing landscape generated via NK model
}
\description{
A fitness landscape generated by the rough Mt. Fuji (RMF) model.
The landscape is returned as a two-column data frame,
where the first column contains the coordinates of
the spot (class `character`) and the second column contains
the fitness value (class `numeric`). For more information
on RMF models, see Aita et al. (2000)
<doi:10.1002/(SICI)1097-0282(200007)54:1<64::AID-BIP70>3.0.CO;2-R>.
}
