% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_sb}
\alias{generate_sb}
\title{Stickbreaking Model for Fitness Landscapes}
\usage{
generate_sb(n_gene, n_allele, wt_fit, mut_fit, max_fit)
}
\arguments{
\item{n_gene}{number of genes}

\item{n_allele}{number of alleles per gene}

\item{wt_fit}{wild-type genotype fitness}

\item{mut_fit}{if of class \code{matrix}, contains numeric fitness values
for each allele and gene combination (each column is a gene, each row is
an allele); if a \code{function}, must have zero parameters and return a
single \code{numeric} value upon each call (it will be called repeatedly
when generating the landscape)}

\item{max_fit}{maximum fitness value attainable in this landscape}
}
\value{
fitness landscape stored in \code{FitLand} class object
}
\description{
Generates a fitness landscape using the stickbreaking model, as
characterized in Nagel et al. (2012) <doi:10.1534.genetics.111.132134>.
}
\examples{
***
}
