% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{getDend.lst}
\alias{getDend.lst}
\title{A function to generate a list of dendrograms with various sd cutoffs}
\usage{
getDend.lst(mat, lbl, vec.sd, dist.method, hclust.method, vec.colors,
  colorFn = labelNode)
}
\arguments{
\item{mat}{matrix of data}

\item{lbl}{data table with sample labels}

\item{vec.sd}{vector of SD cutoffs to use}

\item{dist.method}{distance method}

\item{hclust.method}{hclust method}

\item{vec.colors}{color vector indexed by sample group}

\item{colorFn}{dendrogram coloring function (ex. labelNode)}
}
\value{
a list of dendrograms
}
\description{
A function to generate a list of dendrograms with various sd cutoffs
}
