% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intake_tracker.R
\name{intake_tracker}
\alias{intake_tracker}
\title{Intake Tracker}
\usage{
intake_tracker(c)
}
\arguments{
\item{Positive}{Integer}
}
\value{
Altered 'intakes' vector, new 'days' vector, 'cdf' dataframe, and ggplot2 visualization
}
\description{
Intake Tracker
}
\examples{
intake_tracker(2100)
}
