% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointfun.R
\name{jointfun}
\alias{jointfun}
\title{jointing things is indeed fun!  I need to read around to figure out what this
is doing.}
\usage{
jointfun(q, lambda, lp, sigma, nta)
}
\arguments{
\item{q}{not p}

\item{lambda}{not mu}

\item{lp}{Something which I do not know what it is.}

\item{sigma}{no, I want tau!  0.5 pi makes more sense.}

\item{nta}{The number of TAs.}
}
\value{
vector of lambda * (the bagelfuns() + (1 - labmda) * the low n
  bfuns!)
}
\description{
I want to grumble for a moment; why do science people insist on such stupid
names for variables. Look, I get it that sigma usually means standard
deviation and mu mean; but cannot we get a frigging clue to remind us what is
being counted?
}
