% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{labelNode}
\alias{labelNode}
\title{A helper function to color dendrogram labels}
\usage{
labelNode(x, lbl, vec.colors)
}
\arguments{
\item{x}{dendrogram element}

\item{lbl}{data table of labels}

\item{vec.colors}{vector of colors indexed by sample group}
}
\value{
dendrogram element
}
\description{
A helper function to color dendrogram labels
}
