% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{lstToMatrix.mat}
\alias{lstToMatrix.mat}
\title{List to matrix}
\usage{
lstToMatrix.mat(lst, str.id, str.measure, fn.aggregate = sum)
}
\arguments{
\item{lst}{list of data.tables}

\item{str.id}{string with id variable name}

\item{str.measure}{string with measure variable name}

\item{fn.aggregate}{aggregation function. Default: sum}
}
\value{
a matrix with measure variable as rows, list names as columns, measuring the measure var using fn.aggregate
}
\description{
Function to convert a list of data.tables into a matrix with list names as columns, id
}
