% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_functions.r
\name{modelPairs.dt}
\alias{modelPairs.dt}
\title{ModelPairs}
\usage{
modelPairs.dt(dt, col.id, col.x, col.y, fn.model = lm)
}
\arguments{
\item{dt}{data.table with data}

\item{col.id}{id in data.table (ex. study day)}

\item{col.x}{x variable}

\item{col.y}{y variable (response)}

\item{fn.lstAggregate}{function to aggregate list data}

\item{str.variable}{name for id column in new table after aggregation over all col.id in vec.idRange}

\item{str.value}{name for value column in new table}

\item{fn.rangeAggregate}{function to aggregate data in new table (ex. median of days in the specified date range). Default: median}
}
\value{
data.table with aggregated data
}
\description{
Function to create a model with a pair of variables
}
