% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cvm.R
\name{my_cvm}
\alias{my_cvm}
\title{This is not your mother's Cramer-von Mises goodness of fit function!}
\usage{
my_cvm(x, null = "punif", ..., nullname)
}
\arguments{
\item{x}{the something I do not yet understand.}

\item{null}{The null distribution we are comparing against.}

\item{...}{Extra arguments passed to intersect()}

\item{nullname}{name of the null hypothesis to be printed upon completion.}
}
\value{
List containing the result of the test.
}
\description{
Calculate a goodness of fit metric given something which I do not yet understand.
}
