% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{plotPCA.lst}
\alias{plotPCA.lst}
\title{A function to produce a list of PCA plots}
\usage{
plotPCA.lst(pca.obj, lbl, str.label, vec.colors.type, p.theme,
  col.group = "group")
}
\arguments{
\item{pca.obj}{pca object}

\item{lbl}{data.table with sample labels and groups}

\item{str.label}{string with plot title}

\item{vec.colors.type}{color vector for sample types}

\item{p.theme}{ggplot2 theme variable}

\item{col.group}{group column in lbl. Default: 'group'}
}
\value{
list of plots
}
\description{
A function to produce a list of PCA plots
}
