% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.r
\name{plotScatter.plt}
\alias{plotScatter.plt}
\title{Plot Scatter}
\usage{
plotScatter.plt(dt, col.x, col.y, vec.xlim, vec.ylim, str.title, str.xlab,
  str.ylab, p.theme, num.alpha = 0.07, num.size = 0.5,
  trend.method = "lm")
}
\arguments{
\item{dt}{data.table to plot}

\item{col.x}{x column label}

\item{col.y}{y column label}

\item{vec.xlim}{vector of x limits (min, max)}

\item{vec.ylim}{vector of y limits (min, max)}

\item{str.title}{plot title}

\item{str.xlab}{x axis label}

\item{str.ylab}{y axis label}

\item{p.theme}{plot theme object}

\item{num.alpha}{alpha for dots. Default: 0.07}

\item{num.size}{size for dots. Defaults: 0.5}

\item{trend.method}{method for trendline. Default: 'lm'}
}
\value{
plot object
}
\description{
Generic scatterplot function
}
