% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{read.timeFile.lst}
\alias{read.timeFile.lst}
\title{Read time file}
\usage{
read.timeFile.lst(lbl, col.time = "StartTime",
  str.format = "\%m/\%d/\%Y \%H:\%M:\%S \%p", str.tz = "UTC",
  col.dir = "dir", col.file = "file", col.label = "label")
}
\arguments{
\item{lbl}{sample key table with directory, file name, and label columns}

\item{col.time}{column name or vector of column names with date/time data. Default: 'StartTime'}

\item{str.format}{incoming date format or vector of date formats corresponding to each element of the col.time vector. Default: '%m/%d/%Y %H:%M:%S %p' (1/22/19 12:33:45 PM)}

\item{str.tz}{time zone for date conversion. Default: 'UTC' (avoids double counting issues)}

\item{col.dir}{column name in lbl with directory. Default: 'dir'}

\item{col.file}{column name in lbl with file. Default: 'file'}

\item{col.label}{column name in lbl with label. Default: 'label'}
}
\value{
data.table with time columns in posix format
}
\description{
Reads a file with date/time data using information in a label table
}
