% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataManipulation_functions.r
\name{readTables.lst}
\alias{readTables.lst}
\title{Read tables}
\usage{
readTables.lst(dt.lbl, col.label = "label", col.dir = "dir",
  col.file = "file")
}
\arguments{
\item{dt.lbl}{data.table with sample labels, directory (can be relative to root), and file names}

\item{col.label}{label column name}

\item{col.dir}{directory column name}

\item{col.file}{file column name}
}
\value{
list of data.tables named by label
}
\description{
Function to read in a set of data tables specified by a label table
}
