% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tallyprepfun.R
\name{tallyprepfun}
\alias{tallyprepfun}
\title{Get the set of useful TAs from a tally file, I think.}
\usage{
tallyprepfun(strain, file_location)
}
\arguments{
\item{strain}{Which strain?}

\item{file_location}{Where is the tally file?}
}
\value{
Set of tallies to use.
}
\description{
Get the set of useful TAs from a tally file, I think.
}
