% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_calorie_matrix.R
\name{weight_calorie_matrix}
\alias{weight_calorie_matrix}
\title{Weight and Calorie Visualization Matrix}
\usage{
weight_calorie_matrix()
}
\value{
Side by side ggplot2 visualizations
}
\description{
Weight and Calorie Visualization Matrix
}
