% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_tracker.R
\name{weight_tracker}
\alias{weight_tracker}
\title{Weight Tracker}
\usage{
weight_tracker(w, g)
}
\arguments{
\item{Positive}{Integer (current weight in kg)}

\item{Postive}{Integer (goal weight in kg)}
}
\value{
Altered 'weights' vector, new 'x' vector, 'wdf' dataframe, and ggplot2 visualization
}
\description{
Weight Tracker
}
\examples{
weight_tracker(70,72)
}
