% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_nk}
\alias{generate_nk}
\title{NK Model for Fitness Landscapes}
\usage{
generate_nk(n, k, n_allele = 2, fitness)
}
\arguments{
\item{n}{number of genes}

\item{k}{number of other loci that influence fitness of each gene;
determines landscape ruggedness}

\item{n_allele}{number of alleles per gene}

\item{fitness}{if of class \code{array}, contains \code{numeric} fitness
values for each (\code{k+1})-length genotype; if a \code{function}, takes
a single parameter (\code{numeric} vector of length \code{k+1}) and
returns a single \code{numeric} fitness value}
}
\value{
fitness landscape stored in \code{FitLand} class object
}
\description{
Generates a fitness landscape using the NK model, as characterized in
Kauffman & Weinberger (1989) <doi:10.1016/s0022-5193(89)80019-0>.
}
