% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_rmf}
\alias{generate_rmf}
\title{Rough Mount Fuji Model for Fitness Landscapes}
\usage{
generate_rmf(n_gene, n_allele, fitness, noise)
}
\arguments{
\item{n_gene}{number of genes}

\item{n_allele}{number of alleles per gene}

\item{fitness}{if of class \code{matrix}, contains the deterministic fitness
of each allele-gene combination (each column is a gene, each row is an
allele); if of class \code{function}, must require zero parameters,
return a single \code{numeric} value, and will be repeatedly called for
each genotype to determine deterministic fitness}

\item{noise}{if of class \code{array}, contains the (pseudo)random
contribution to each genotype; if of class \code{function}, must require
zero parameters, return a single \code{numeric} value, and will be
repeatedly called for each genotype to determine (pseudo)random fitness
contribution}
}
\value{
fitness landscape stored in \code{FitLand} class object
}
\description{
Generates a fitness landscape using the rough Mt. Fuji model, as
characterized in Aita et al. (2000)
<doi:10.1002/(SICI)1097-0282(200007)54:1<64::AID-BIP70>3.0.CO;2-R>.
}
\examples{
num_genes <- 4
num_alelles <- 3
landscape <- generate_rmf(num_genes,
                          num_alleles,
                          fitness = function(){stats::rnorm(1, mean = 25)},
                          noise = function(){stats::rnorm(1, mean = 2)})
                          
print(landscape)
}
