% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_raw_tally.R
\name{import_raw_tally}
\alias{import_raw_tally}
\title{Read in some raw tally files.}
\usage{
import_raw_tally(filename, CIA = TRUE)
}
\arguments{
\item{filename}{File to parse.}

\item{CIA}{I have no clue.}
}
\value{
data frame of the tally information.
}
\description{
Read in some raw tally files.
}
